+++ src/main/java/io/github/forgestove/create_cyber_goggles/event/MouseScroll.java
@@ -5,7 +5,7 @@ import net.minecraftforge.client.event.InputEvent.MouseScrollingEvent;
 public class MouseScroll {
 	public static void onMouseScroll(MouseScrollingEvent event) {
 		if (!CCG.CONFIG.goggles.betterStoreInfo) return;
-		if (Common.getSelectedBE() instanceof TableClothBlockEntity tcbe && tcbe.isShop()) {
+		if (Common.getBE() instanceof TableClothBlockEntity tcbe && tcbe.isShop()) {
 			if (event.getScrollDelta() == 0) Common.scrollDeltaY = 0;
 			else Common.scrollDeltaY = event.getScrollDelta() > 0 ? -1 : 1;
 			event.setCanceled(true);