+++ CHANGELOG.md
@@ -1,1 +1,7 @@
-- Fixed ConcurrentModificationException caused by delay render
+- Added outline for package port
+- Added tooltip for blaze burner
+- Optimized interact feedback for hotkey
+- Optimized kinetic debugger for chain conveyor
+- Optimized localization
+- Refactor key input
+- Removed support for de_de, ru_ru, zh_tw